+++ src/include/at.h
@@ -25,9 +25,9 @@ struct At{
     At_Err_t (*handleAuto)(struct At* this);
 
     // size_t (*printf)(struct At* this, const char* format, ...)  __attribute__ ((format (printf, 2, 3)));
-    size_t (*printf)(struct At* this, const char* format, ...);
-    size_t (*print)(struct At* this, const char* message);
-    size_t (*println)(struct At* this, const char* message);
+    int (*printf)(struct At* this, const char* format, ...);
+    int (*print)(struct At* this, const char* message);
+    int (*println)(struct At* this, const char* message);
     At_Err_t (*printSet)(struct At* this, const char* name);
     At_Err_t (*sendInfor)(struct At* this, const char* infor);
